+++ src/form-elements/FormElementABN.tsx
@@ -120,7 +120,8 @@ function FormElementABN({
         })
         if (result.ABRPayloadSearchResults.response.exception) {
           throw new Error(
-            result.ABRPayloadSearchResults.response.exception.exceptionDescription,
+            result.ABRPayloadSearchResults.response.exception
+              .exceptionDescription,
           )
         }
 
@@ -269,6 +270,7 @@ function FormElementABN({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>
         {value && (