+++ src/form-elements/FormElementCheckBoxes.tsx
@@ -213,6 +213,7 @@ function FormElementCheckboxes({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </FormElementOptions>